Refactor to generic name for card
